+++ sdk/ml/azure-ai-ml/tests/automl_job/unittests/test_automl_image_classification.py
@@ -4,8 +4,8 @@
 
 import pytest
 
-from azure.ai.ml import UserIdentity
-from azure.ai.ml._restclient.v2022_06_01_preview.models import (
+from azure.ai.ml import UserIdentityConfiguration
+from azure.ai.ml._restclient.v2022_10_01_preview.models import (
     ClassificationPrimaryMetrics,
     ImageModelSettingsClassification,
     LearningRateScheduler,
@@ -13,7 +13,7 @@ from azure.ai.ml._restclient.v2022_06_01_preview.models import (
     SamplingAlgorithmType,
     StochasticOptimizer,
 )
-from azure.ai.ml._restclient.v2022_06_01_preview.models import UserIdentity as RestUserIdentity
+from azure.ai.ml._restclient.v2022_10_01_preview.models import UserIdentity as RestUserIdentity
 from azure.ai.ml.automl import image_classification
 from azure.ai.ml.constants._common import AssetTypes
 from azure.ai.ml.entities._inputs_outputs import Input
@@ -26,7 +26,7 @@ from azure.ai.ml.sweep import BanditPolicy, Choice, Uniform
 class TestAutoMLImageClassification:
     @pytest.mark.parametrize("run_type", ["single", "sweep", "automode"])
     def test_image_classification_task(self, run_type):
-        identity = UserIdentity()
+        identity = UserIdentityConfiguration()
         # Create AutoML Image Classification task
         image_classification_job = image_classification(
             training_data=Input(type=AssetTypes.MLTABLE, path="https://foo/bar/train.csv"),