+++ sdk/ml/azure-ai-ml/tests/workspace/e2etests/test_workspace.py
@@ -11,6 +11,7 @@ from azure.ai.ml.constants._common import PublicNetworkAccess
 from azure.ai.ml.entities._workspace.diagnose import DiagnoseResponseResultValue
 from azure.ai.ml.entities._workspace.identity import ManagedServiceIdentityType
 from azure.ai.ml.entities._workspace.workspace import Workspace
+from azure.ai.ml.constants._workspace import ManagedServiceIdentityType
 from azure.core.paging import ItemPaged
 from azure.core.polling import LROPoller
 from azure.mgmt.msi._managed_service_identity_client import ManagedServiceIdentityClient