+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -14,6 +14,12 @@
 - Deprecated parameter path has been removed from load and dump methods.
 - JobOperations.cancel() is renamed to JobOperations.begin_cancel() and it returns LROPoller
 - Workspace.list_keys renamed to Workspace.get_keys.
+- OnlineDeploymentOperations.delete has been renamed to begin_attach.
+- Datastore credentials are switched to use unified credential configuration classes.
+- UserAssignedIdentity is replaced by ManagedIdentityConfiguration
+- Job using unified identity classes.
+- Endpoint uses unified identity classes.
+- Workspace ManagedServiceIdentity has been replaced by IdentityConfiguration.
 
 ### Bugs Fixed
 - Fix identity passthrough job with single file code
@@ -22,6 +28,7 @@
  - Removed declaration on Python 3.6 support
  - Added support for custom setup scripts on compute instances.
  - Updated dependencies upper bounds to be major versions.
+ - Switched Compute operations to use Oct preview API version.
 
 ## 0.1.0b7 (2022-09-22)